refactor: replace antd Button in FieldsForm (Workspace & Project Settings)
